primary ray tracing

+++ main.h
@@ -29,7 +29,7 @@ struct world
 internal void write_ppm(f32 *data, int width, int height, f32 gamma, char *fname);
 internal pixel_buffer new_pixel_buffer(u32 width, u32 height);
 internal void clear(pixel_buffer buffer, v4 c);
-internal void set_pixel(pixel_buffer buffer, f32 x, f32 y, v4 color);
+internal void set_pixel(pixel_buffer buffer, int x, int y, v4 color);
 
 inline void
 set_gamma(view_plane *vp, f32 gamma)
